adding test to a heavy user time program
