+++ Makefile
@@ -168,7 +168,7 @@ mkfs: mkfs.c fs.h
 UPROGS=\
 	_cat\
 	_time\
-	_frog\
+	_heavyUser\
 	_echo\
 	_forktest\
 	_grep\
